refactor: split large modules
