+++ crates/cli/src/exec/mod.rs
@@ -1,0 +1,5 @@
+pub(crate) mod privileges;
+pub(crate) mod transfer;
+
+pub(crate) use privileges::check_privileges;
+pub(crate) use transfer::execute_transfer;
